backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
jwcrypto==1.5.6
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
